+++ backend/src/main/java/dev/wstein/flixplugin/FlixRunMainAction.java
@@ -48,10 +48,7 @@ public class FlixRunMainAction extends LSPCommandAction {
         Path jar = FlixFork.resolveJar(project);
         String entryPoint = entryPointOf(command.getArguments());
 
-        GeneralCommandLine commandLine = new GeneralCommandLine("java", "-jar", jar.toString(), "run");
-        if (entryPoint != null) {
-            commandLine.addParameters("--entrypoint", entryPoint);
-        }
+        GeneralCommandLine commandLine = new GeneralCommandLine(FlixLaunchCommand.run(jar, entryPoint));
         String basePath = project.getBasePath();
         if (basePath != null) {
             commandLine.setWorkDirectory(basePath);